requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
dj-database-url==0.4.2
django-localflavor==1.5.1
django==1.11.2
gunicorn==19.7.1
psycopg2==2.7.1
pytz==2017.2              # via django
whitenoise==3.3.0
